CLICK THE SUBMIT BUTTON
In order to make this project as real as possible, I made the "Sbmit" button into a hyperlink (to make the end-user feel like they just inquired about the space station).

I am having difficulty with my animated vids and gifs: I want to use them for a backgraound; but they will not load.